backend/requirements.txt
Flask
easyocr
Pillow
pandas
roboflow
opencv-python-headless

